fix: Include all changes since last regular release only in GitHub releases.
When doing regular releases include all the changes since the last
regular release - including changes from pre-releases in the release
notes on GitHub, but not in the CHANGES.md file. In the CHANGES.md file
every change can be seen at once. Including all the changes since the
previous regular release is useful for the GitHub release page only.